Ju/news_database/requirements.txt
# Airflow (MySQL 연동 포함) + 안정된 의존성
apache-airflow[mysql]==2.9.3

# 웹 크롤링 / 데이터 처리
selenium==4.23.1
beautifulsoup4==4.12.3
pandas==2.1.4

# ChromeDriver (설치된 Chrome 버전에 맞게 조정)
chromedriver-binary==140.0.7322.0.0